add transactions to repositories
Now library has a flag to set it transactional and is used with all
locations and repositories below it. Transaction is managed from the
location with Commit and Rollback methods. While a transaction is being
made no other transaction can start.

When a transaction starts a new siva filesystem is created for it and
the previous cachedFS kept. This way previously created repositories
from that location should work the same as before. It should also be
able to create new repositories.

Signed-off-by: Javi Fontan <jfontan@gmail.com>